Show ordered and onStack on ComponentList Create ComponentPurchaseList.vue

+++ spring/src/main/java/com/noovitec/mpb/dto/InventoryDto.java
@@ -14,6 +14,6 @@ public class InventoryDto {
 	private String componentName;
 	private String supplierName;
 	private int unitsOnStack;
-	private Long unitsOrdered;
+	private int unitsOrdered;
 
 }
